// src/models/kmap.ts
export interface MenuUser {
    GROUP_ID: string;
    GROUP_NAME: string;
    TOOL_ID: string;
    TOOL_NAME: string;
    URL: string;
    ICON_SRC: string;
    TOOL_ORDER: string;
    genId: string;
}


export interface CategoryMenuUser {
    GROUP_ID: string;
    GROUP_NAME: string;
    genId: string;
}